Move index data classes to a new file
