+++ tests/test_text/test_text_comments.py
@@ -1,9 +1,9 @@
 """
-novelWriter – NWIndex Class Tester
-==================================
+novelWriter – Text Comment Tester
+=================================
 
 This file is a part of novelWriter
-Copyright (C) 2020 Veronica Berglyd Olsen and novelWriter contributors
+Copyright (C) 2025 Veronica Berglyd Olsen and novelWriter contributors
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
